images/pythonruntime/nimbella-sdk-python/requirements.txt
## Both of these packages are installed by the runtime already
google-cloud-storage == 1.28.1
redis == 3.4.1
